requirements.txt
Django==2.1.1
pytz==2018.5
httplib2==0.11.3
feedparser==5.2.1
Markdown==2.6.11
psycopg2==2.7.5
Pillow==5.2.0
versiontools==1.9.1
statsd==3.3.0
pep8==1.7.1
pyflakes==2.0.0
mccabe==0.6.1
configparser==3.5.0
pycodestyle==2.3.1 # pyup: <2.4.0
flake8==3.5.0
tornado==5.1
beautifulsoup4==4.6.3
cssselect==1.0.3
lxml==4.2.4
fuzzywuzzy==0.17.0
sure==1.4.11
traitlets==4.3.2  # ipython
ipython_genutils==0.2.0  # ipython
decorator==4.2.1  # ipython
ptyprocess==0.6.0  # ipython
pexpect==4.6.0  # ipython
pathlib2==2.3.2  # ipython
pickleshare==0.7.4  # ipython
simplegeneric==0.8.1  # ipython
path.py==11.1.0  # ipython
backports.shutil_get_terminal_size==1.0.0  # ipython
wcwidth==0.1.7  # ipython
prompt_toolkit==2.0.4  # ipython
pygments==2.2.0  # ipython
scandir==1.9.0  # ipython
ipython==6.5.0
ipdb==0.11
rdflib==4.2.2
coverage==4.5.1
pyasn1==0.4.4
cryptography==2.3.1  # pyOpenSSL
pyOpenSSL==18.0.0
ndg-httpsclient==0.5.1
urllib3==1.22  # requests
chardet==3.0.4  # requests
idna==2.7  # requests
certifi==2018.8.24  # requests
requests==2.19.1
logilab-common==1.4.2
logilab-astng==0.24.3
astroid==2.0.4
pylint==2.1.1
six==1.11.0
ipaddress==1.0.21  # faker
faker==0.8.18  # factory_boy
factory_boy==2.11.1
ldap3==2.5.1
sqlparse==0.2.4
boto==2.49.0
python-dateutil==2.7.3

djangowind==1.2.1
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.10.4
django-appconf==1.0.2
rcssmin==1.0.6  # django-compressor
rjsmin==1.0.12  # django-compressor
django-compressor==2.2
django-statsd-mozilla==0.4.0
raven==6.9.0
django-bootstrap4==0.0.7
django-debug-toolbar==1.10.1
django-waffle==0.14.0
django-jenkins==0.110.0
django-smoketest==1.1.0
typing==3.6.6  # For django-extensions
django-extensions==2.1.2
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==1.3
django-markwhat==1.6.1
gunicorn==19.8.1
django-infranil==1.1.0
django-flatblocks==0.9.4
django-storages-redux==1.3.3

djangorestframework==3.8.2

ccnmtlsettings==1.4.0

django-reversion==3.0.0
text_unidecode==1.2

django-taggit==0.23.0

django-cacheds3storage==0.1.2

nameparser==1.0.1
oauth2==1.9.0.post1
oauthlib==2.1.0
pylti==0.7.0
django-lti-provider==0.3.3

pyparsing==2.2.0
edtf==3.0.0 # pyup: <4.0.0

pbr==4.2.0 # bandit
PyYAML==3.13 # bandit
stevedore==1.29.0  # bandit
bandit==1.5.1

django-smtp-ssl==1.0
django-contact-us==0.4.1